+++ CHANGELOG.rst
@@ -4,6 +4,12 @@ Changelog
 `Unreleased <https://github.com/staticjinja/staticjinja/compare/main>`_
 -----------------------------------------------------------------------
 
+Changed
+^^^^^^^
+
+* Stop testing for Python 3.6, add testing for Python 3.10. See
+  https://github.com/staticjinja/staticjinja/pull/174
+
 `4.1.3 <https://github.com/staticjinja/staticjinja/compare/4.1.2...4.1.3>`_ (2022-06-03)
 ----------------------------------------------------------------------------------------