+++ CHANGELOG.rst
@@ -17,6 +17,9 @@ Removed
 * Remove previously deprecated ``Site.get_dependencies()``.
   Use ``Site.get_dependents()`` instead. See
   https://github.com/staticjinja/staticjinja/commit/170b027a4fff86790bc69a1222d7b0a36c1080bc
+* Removed previously deprecated ``logger`` argument to Site.__init__().
+  Use ``staticjinja.logger`` for the staticjinja-wide logger instead. See
+  https://github.com/staticjinja/staticjinja/issues/144
 * Stop testing for Python 3.6. It still may work, just not officially supported.
 
 Added